Refactored utility contexts out of context
