+++ src/lib.rs
@@ -5,3 +5,5 @@ pub mod templates;
 pub mod types;
 pub use dam_core::*;
 pub use dam_macros::*;
+
+pub mod utility_contexts;
