+++ src/utility_contexts/approx_checker_context.rs
@@ -1,27 +1,30 @@
-use dam_core::prelude::*;
 use dam_macros::context;
 
+use crate::context::Context;
 use crate::{
     channel::{ChannelElement, DequeueResult, Receiver},
     types::DAMType,
 };
-
-use super::Context;
+use dam_core::prelude::*;
 
 #[context]
-pub struct CheckerContext<T: Clone, IType, FType>
+pub struct ApproxCheckerContext<T: Clone, IType, FType, CheckType>
 where
     IType: Iterator<Item = T>,
     FType: FnOnce() -> IType + Send + Sync,
+    CheckType: Send + Sync,
 {
     iterator: Option<FType>,
     input: Receiver<T>,
+    checker: CheckType,
 }
 
-impl<T: DAMType, IType, FType> Context for CheckerContext<T, IType, FType>
+impl<T: DAMType, IType, FType, CheckType> Context
+    for ApproxCheckerContext<T, IType, FType, CheckType>
 where
     IType: Iterator<Item = T>,
     FType: FnOnce() -> IType + Send + Sync,
+    CheckType: Fn(&T, &T) -> bool + Send + Sync,
 {
     fn init(&mut self) {}
 
@@ -29,7 +32,9 @@ where
         if let Some(iter) = self.iterator.take() {
             for (ind, val) in iter().enumerate() {
                 match self.input.dequeue(&self.time) {
-                    DequeueResult::Something(ChannelElement { time, data }) if data != val => {
+                    DequeueResult::Something(ChannelElement { time, data })
+                        if !(self.checker)(&val, &data) =>
+                    {
                         panic!("Mismatch on iteration {ind} at time {time:?}: Expected {val:?} but found {data:?}")
                     }
                     DequeueResult::Something(_) => {}
@@ -44,15 +49,17 @@ where
     }
 }
 
-impl<T: DAMType, IType, FType> CheckerContext<T, IType, FType>
+impl<T: DAMType, IType, FType, CheckType> ApproxCheckerContext<T, IType, FType, CheckType>
 where
     IType: Iterator<Item = T>,
     FType: FnOnce() -> IType + Send + Sync,
+    CheckType: Fn(&T, &T) -> bool + Send + Sync,
 {
-    pub fn new(iterator: FType, input: Receiver<T>) -> CheckerContext<T, IType, FType> {
-        let gc = CheckerContext {
+    pub fn new(iterator: FType, input: Receiver<T>, checker: CheckType) -> Self {
+        let gc = Self {
             iterator: Some(iterator),
             input,
+            checker,
             context_info: Default::default(),
         };
         gc.input.attach_receiver(&gc);
