+++ src/utility_contexts/checker_context.rs
@@ -1,30 +1,27 @@
+use dam_core::prelude::*;
 use dam_macros::context;
 
-use super::Context;
 use crate::{
     channel::{ChannelElement, DequeueResult, Receiver},
     types::DAMType,
 };
-use dam_core::prelude::*;
+
+use crate::context::Context;
 
 #[context]
-pub struct ApproxCheckerContext<T: Clone, IType, FType, CheckType>
+pub struct CheckerContext<T: Clone, IType, FType>
 where
     IType: Iterator<Item = T>,
     FType: FnOnce() -> IType + Send + Sync,
-    CheckType: Send + Sync,
 {
     iterator: Option<FType>,
     input: Receiver<T>,
-    checker: CheckType,
 }
 
-impl<T: DAMType, IType, FType, CheckType> Context
-    for ApproxCheckerContext<T, IType, FType, CheckType>
+impl<T: DAMType, IType, FType> Context for CheckerContext<T, IType, FType>
 where
     IType: Iterator<Item = T>,
     FType: FnOnce() -> IType + Send + Sync,
-    CheckType: Fn(&T, &T) -> bool + Send + Sync,
 {
     fn init(&mut self) {}
 
@@ -32,9 +29,7 @@ where
         if let Some(iter) = self.iterator.take() {
             for (ind, val) in iter().enumerate() {
                 match self.input.dequeue(&self.time) {
-                    DequeueResult::Something(ChannelElement { time, data })
-                        if !(self.checker)(&val, &data) =>
-                    {
+                    DequeueResult::Something(ChannelElement { time, data }) if data != val => {
                         panic!("Mismatch on iteration {ind} at time {time:?}: Expected {val:?} but found {data:?}")
                     }
                     DequeueResult::Something(_) => {}
@@ -49,17 +44,15 @@ where
     }
 }
 
-impl<T: DAMType, IType, FType, CheckType> ApproxCheckerContext<T, IType, FType, CheckType>
+impl<T: DAMType, IType, FType> CheckerContext<T, IType, FType>
 where
     IType: Iterator<Item = T>,
     FType: FnOnce() -> IType + Send + Sync,
-    CheckType: Fn(&T, &T) -> bool + Send + Sync,
 {
-    pub fn new(iterator: FType, input: Receiver<T>, checker: CheckType) -> Self {
-        let gc = Self {
+    pub fn new(iterator: FType, input: Receiver<T>) -> CheckerContext<T, IType, FType> {
+        let gc = CheckerContext {
             iterator: Some(iterator),
             input,
-            checker,
             context_info: Default::default(),
         };
         gc.input.attach_receiver(&gc);
